requirements.txt
streamlit==1.38.0
streamlit-webrtc==0.47.1
openai==1.43.0
python-dotenv==1.0.1
PyPDF2==3.0.1
tiktoken==0.7.0
chromadb==0.5.5
sentence-transformers==3.0.1
scikit-learn==1.5.1
numpy==1.26.4
